requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.60.0
tensorboard>=2.8.0

# Optional dependencies
# lpips>=0.1.4  # Uncomment for LPIPS metric